requirements.txt
opentimelineio==0.14.0
ffmpeg-python==0.2.0
